docs: simplify site generation and correct CLI references
