+++ docs/development.html
@@ -276,6 +276,7 @@ body:not(.home) .doc>h1:first-child{display:none}
 <p>The Markdown sources and generated GitHub Pages site live in <code>docs/</code>.</p>
 <pre><code class="language-bash"><span class="hl-cmd">npm</span> install
 <span class="hl-cmd">make</span> docs-site</code></pre>
+<p>The builder generates HTML and <code>llms.txt</code> together and validates local links. Commit the regenerated files with Markdown changes. The Markdown renderer, syntax highlighter, and shared HTML escaping live in separate <code>scripts/docs-*.mjs</code> modules.</p>
 <h2 id="ghostty-web-snapshot"><a class="anchor" href="#ghostty-web-snapshot" aria-label="Anchor link">#</a>Ghostty web snapshot</h2>
 <p>Install Playwright&#39;s Chromium build once, then run the snapshot target:</p>
 <pre><code class="language-bash">npx playwright install chromium
