+++ docs/json.html
@@ -291,8 +291,8 @@ body:not(.home) .doc>h1:first-child{display:none}
   }
 ]</code></pre>
 <h2 id="fields"><a class="anchor" href="#fields" aria-label="Anchor link">#</a>Fields</h2>
-<table><thead><tr><th>Field</th><th>Type</th><th>Notes</th></tr></thead><tbody><tr><td><code>id</code></td><td>string</td><td>Provider&#39;s stable identifier.</td></tr><tr><td><code>title</code></td><td>string</td><td>Human title; falls back to provider description, then <code>id</code>.</td></tr><tr><td><code>url</code></td><td>string</td><td>Direct GIF URL — what <code>--format url</code> prints.</td></tr><tr><td><code>preview_url</code></td><td>string</td><td>Smaller GIF good for thumbnails / inline previews.</td></tr><tr><td><code>tags</code></td><td>string[]</td><td>May be empty depending on provider.</td></tr><tr><td><code>width</code></td><td>number</td><td>Original GIF width in pixels (0 if unknown).</td></tr><tr><td><code>height</code></td><td>number</td><td>Original GIF height in pixels (0 if unknown).</td></tr></tbody></table>
-<p>The shape is identical across providers. Field values may differ — KLIPY tends to ship tags more aggressively than GIPHY, and dimensions can be <code>0</code> when the upstream omits them.</p>
+<table><thead><tr><th>Field</th><th>Type</th><th>Notes</th></tr></thead><tbody><tr><td><code>id</code></td><td>string</td><td>Provider&#39;s stable identifier.</td></tr><tr><td><code>title</code></td><td>string</td><td>Human title; falls back to provider description, then <code>id</code>.</td></tr><tr><td><code>url</code></td><td>string</td><td>Direct GIF URL — what <code>--format url</code> prints.</td></tr><tr><td><code>preview_url</code></td><td>string</td><td>Smaller GIF good for thumbnails / inline previews.</td></tr><tr><td><code>tags</code></td><td>string[]</td><td>Omitted when empty; supplied by KLIPY when available.</td></tr><tr><td><code>width</code></td><td>number</td><td>Original GIF width in pixels; omitted when zero/unknown.</td></tr><tr><td><code>height</code></td><td>number</td><td>Original GIF height in pixels; omitted when zero/unknown.</td></tr></tbody></table>
+<p>Both providers use this shape. GIPHY results omit tags; either provider can omit unknown dimensions.</p>
 <h2 id="stability"><a class="anchor" href="#stability" aria-label="Anchor link">#</a>Stability</h2>
 <ul>
 <li>New optional fields may be added without warning.</li>
