+++ docs/providers/giphy.html
@@ -292,7 +292,7 @@ body:not(.home) .doc>h1:first-child{display:none}
 <li><code>url</code> ← <code>images.original.url</code></li>
 <li><code>preview_url</code> ← <code>images.fixed_width_small.url</code></li>
 <li><code>width</code> / <code>height</code> ← <code>images.original.width</code>/<code>.height</code></li>
-<li><code>tags</code> ← derived from title tokens (GIPHY doesn&#39;t expose a native tag list on search)</li>
+<li><code>tags</code> is omitted (GIPHY doesn&#39;t expose a native tag list on search).</li>
 </ul>
 <h2 id="limits-and-ratings"><a class="anchor" href="#limits-and-ratings" aria-label="Anchor link">#</a>Limits and ratings</h2>
 <ul>
