+++ docs/quickstart.html
@@ -292,7 +292,7 @@ body:not(.home) .doc>h1:first-child{display:none}
 <li><code>/</code> — edit search</li>
 <li><code>↑</code> / <code>↓</code> — select</li>
 <li><code>d</code> — download to <code>~/Downloads</code></li>
-<li><code>f</code> — reveal last download in your file manager</li>
+<li><code>f</code> — download if needed and reveal the selected GIF in your file manager</li>
 <li><code>q</code> — quit</li>
 </ul>
 <p>In Kitty, Ghostty, or iTerm2 you&#39;ll get <strong>animated inline previews</strong>. See <a href="previews.html">Previews</a> for protocol details.</p>
