+++ docs/tui.html
@@ -274,13 +274,13 @@ body:not(.home) .doc>h1:first-child{display:none}
 <span class="hl-cmd">gifgrep</span> tui <span class="hl-f">--source</span> giphy cats</code></pre>
 <p><img src="assets/gifgrep-tui.png" alt="gifgrep TUI with animated inline previews"></p>
 <h2 id="keys"><a class="anchor" href="#keys" aria-label="Anchor link">#</a>Keys</h2>
-<table><thead><tr><th>Key</th><th>Action</th></tr></thead><tbody><tr><td><code>/</code></td><td>Edit search query.</td></tr><tr><td><code>↑</code> <code>↓</code></td><td>Move selection.</td></tr><tr><td><code>d</code></td><td>Download current selection to <code>~/Downloads</code>.</td></tr><tr><td><code>c</code></td><td>Copy the selected GIF to the clipboard.</td></tr><tr><td><code>f</code></td><td>Reveal last download in Finder / Explorer / Files.</td></tr><tr><td><code>q</code></td><td>Quit while browsing; type <code>q</code> while editing a query.</td></tr><tr><td><code>Ctrl-C</code></td><td>Quit from either mode.</td></tr></tbody></table>
+<table><thead><tr><th>Key</th><th>Action</th></tr></thead><tbody><tr><td><code>/</code></td><td>Edit search query.</td></tr><tr><td><code>↑</code> <code>↓</code></td><td>Move selection.</td></tr><tr><td><code>d</code></td><td>Download current selection to <code>~/Downloads</code>.</td></tr><tr><td><code>c</code></td><td>Copy the selected GIF to the clipboard.</td></tr><tr><td><code>f</code></td><td>Download if needed, then reveal the selected GIF.</td></tr><tr><td><code>q</code></td><td>Quit while browsing; type <code>q</code> while editing a query.</td></tr><tr><td><code>Ctrl-C</code></td><td>Quit from either mode.</td></tr></tbody></table>
 <p>Search editing accepts Unicode text, including accented letters, CJK characters, and emoji. Backspace removes one Unicode code point at a time without corrupting the query.</p>
 <p>On Linux, clipboard copy prefers <code>wl-copy</code> in Wayland sessions (<code>WAYLAND_DISPLAY</code> or <code>XDG_SESSION_TYPE=wayland</code>) and <code>xclip</code> otherwise. Install <code>wl-clipboard</code> for Wayland or <code>xclip</code> for X11; if the preferred tool is missing, gifgrep uses the other installed tool. macOS uses its built-in clipboard integration.</p>
 <p>Add <code>--cache</code> (or set <code>GIFGREP_CACHE=1</code>) to reuse explicit downloads across sessions, including GIFs saved by the CLI. The <code>d</code>/<code>f</code> actions still save to <code>~/Downloads</code>. Cache location, age, and size flags match the <a href="search.html#download-cache">download cache options</a>; previews and prefetches remain temporary.</p>
 <h2 id="inline-previews"><a class="anchor" href="#inline-previews" aria-label="Anchor link">#</a>Inline previews</h2>
 <p>The TUI streams animated previews using the <a href="previews.html#kitty-graphics">Kitty graphics protocol</a> on Kitty/Ghostty, <a href="previews.html#iterm2-osc-1337">OSC 1337</a> on iTerm2, Sixel where available, or a truecolor ANSI fallback when forced.</p>
-<table><thead><tr><th>Terminal</th><th>Animation playback</th></tr></thead><tbody><tr><td>Kitty</td><td>Native. gifgrep uploads frames once; the terminal animates them.</td></tr><tr><td>Ghostty</td><td>Software playback (gifgrep ticks frames on a timer).</td></tr><tr><td>iTerm2</td><td>Native — iTerm2 plays GIFs from raw bytes.</td></tr><tr><td>Sixel</td><td>Software playback (gifgrep redraws frames into the same cell area).</td></tr><tr><td>ANSI fallback</td><td>Software playback with truecolor half-blocks.</td></tr><tr><td>Apple Terminal</td><td>Falls back to a static list (no graphics protocol).</td></tr><tr><td>tmux</td><td>Best with a graphics-aware host terminal; YMMV with image passthrough.</td></tr></tbody></table>
+<table><thead><tr><th>Terminal</th><th>Animation playback</th></tr></thead><tbody><tr><td>Kitty</td><td>Native. gifgrep uploads frames once; the terminal animates them.</td></tr><tr><td>Ghostty</td><td>Software playback (gifgrep ticks frames on a timer).</td></tr><tr><td>iTerm2</td><td>Native — iTerm2 plays GIFs from raw bytes.</td></tr><tr><td>Sixel</td><td>Software playback (gifgrep redraws frames into the same cell area).</td></tr><tr><td>ANSI fallback</td><td>Software playback with truecolor half-blocks.</td></tr><tr><td>Apple Terminal</td><td>No graphics detected; use <code>GIFGREP_INLINE=ansi</code> to force truecolor.</td></tr><tr><td>tmux</td><td>Best with a graphics-aware host terminal; YMMV with image passthrough.</td></tr></tbody></table>
 <p>Force or disable software playback:</p>
 <pre><code class="language-bash">GIFGREP_SOFTWARE_ANIM=<span class="hl-n">1</span> <span class="hl-cmd">gifgrep</span> tui cats  <span class="hl-c"> # always tick frames in software</span>
 GIFGREP_SOFTWARE_ANIM=<span class="hl-n">0</span> <span class="hl-cmd">gifgrep</span> tui cats  <span class="hl-c"> # always trust terminal animation</span></code></pre>
@@ -295,7 +295,7 @@ GIFGREP_CELL_ASPECT=<span class="hl-n">0.45</span> <span class="hl-cmd">gifgrep<
 <p>See <a href="providers/">Providers</a> for the full matrix.</p>
 <h2 id="when-the-tui-is-the-wrong-tool"><a class="anchor" href="#when-the-tui-is-the-wrong-tool" aria-label="Anchor link">#</a>When the TUI is the wrong tool</h2>
 <ul>
-<li>For pipelines and scripts, use <a href="search.html"><code>gifgrep search</code></a>. The TUI never writes to stdout.</li>
+<li>For pipelines and scripts, use <a href="search.html"><code>gifgrep search</code></a>. The TUI writes terminal control sequences to stdout.</li>
 <li>For local frame extraction, use <a href="still.html"><code>still</code></a> or <a href="sheet.html"><code>sheet</code></a>. The TUI is search-first.</li>
 <li>In a non-interactive shell (CI, cron), the TUI will refuse to start. Use <code>search</code> with <code>--json</code>.</li>
 </ul><nav class="page-nav" aria-label="Pager"><a class="page-nav-prev" href="search.html"><small>Previous</small><span>Search</span></a><a class="page-nav-next" href="still.html"><small>Next</small><span>still</span></a></nav></article>
