+++ scripts/docs-html.mjs
@@ -1,0 +1,8 @@
+export function escapeHtml(value) {
+  return String(value ?? "").replace(/[&<>"']/g, (char) => ({ "&": "&amp;", "<": "&lt;", ">": "&gt;", '"': "&quot;", "'": "&#39;" })[char]);
+}
+
+export function escapeAttr(value) {
+  return escapeHtml(value);
+}
+
